requirements.txt
alabaster==0.7.12
Babel==2.7.0
certifi==2019.9.11
chardet==3.0.4
coloredlogs==10.0
cycler==0.10.0
docutils==0.15.2
h5py==2.10.0
humanfriendly==4.18
idna==2.8
imagesize==1.1.0
Jinja2==2.10.1
joblib==0.14.0
kiwisolver==1.0.1
m2r==0.2.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mistune==0.8.4
-e git://github.com/mne-tools/mne-python.git@67d03cc0d0554b0d939ddd8ec2d2665add84d4ed#egg=mne
-e git://github.com/jsosulski/moabb.git@914ca448bb4f5051ca3eef823b819bc902d609f6#egg=moabb
numpy==1.16.2
numpydoc==0.9.1
packaging==19.2
pandas==0.24.1
Pillow==6.2.1
Pygments==2.4.2
pyparsing==2.3.1
pyriemann==0.2.5
python-dateutil==2.8.0
pytz==2018.9
PyYAML==5.1.2
requests==2.22.0
scikit-learn==0.21.3
scipy==1.2.1
seaborn==0.9.1
six==1.12.0
sklearn==0.0
snowballstemmer==2.0.0
Sphinx==2.2.1
sphinx-bootstrap-theme==0.7.1
sphinx-gallery==0.4.0
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
statsmodels==0.11.1
urllib3==1.25.7
xgboost==0.90
pytest==5.3.1
